fix: adjustments on attributes types and names

+++ Back/app/schemas/buyer.py
@@ -1,11 +1,12 @@
 from pydantic import BaseModel
 
+
 class BuyerBase(BaseModel):
     name: str
     surname: str
     email: str
     address: str
-    eco_points: int
+    eco_points: float
     dni: str
     billing_address: str
     payment_method: str
@@ -14,6 +15,7 @@ class BuyerBase(BaseModel):
 class BuyerCreate(BuyerBase):
     password: str
     pass
+
 
 class Buyer(BuyerBase):
     id: int
